moves terminator types to sub module
